src/main/java/ru/netology/Radio.java
package ru.netology;

public class Radio {

    protected int currentVolume;
    protected int maxVolume = 100;
    protected int minVolume = 0;
    private int maxStation = 10;
    private int currentStation;

    public Radio() {
        maxStation = 10;
    }

    public Radio(int stationCount) {
        maxStation = stationCount - 1;
    }

    public int getCurrentStation() {
        return currentStation;
    }

    public void setCurrentStation(int newCurrentStation) {
//        if (newCurrentStation < 0) {
//            return;
//        }
        if (newCurrentStation > maxStation) {
            return;
        }
        currentStation = newCurrentStation;
    }

    public int next() {
        if (currentStation == maxStation) {
            currentStation = 0;
        } else {
            currentStation = currentStation + 1;
        }
        return currentStation;
    }


    public int prev() {
        if (currentStation == 0) {
            currentStation = maxStation;
        } else {
            currentStation = currentStation - 1;
        }
        return currentStation;
    }

    public int up() {
        if (currentVolume < maxVolume) {
            currentVolume = currentVolume + 1;
        }
        return currentVolume;
    }

    public int down() {
        if (currentVolume > minVolume) {
            currentVolume = currentVolume - 1;
        }
        return currentVolume;
    }
}




